+++ deeplearning4j/deeplearning4j-scaleout/spark/dl4j-spark/src/test/java/org/deeplearning4j/spark/impl/paramavg/TestSparkMultiLayerParameterAveraging.java
@@ -1,22 +1,27 @@
-/*******************************************************************************
- * Copyright (c) 2015-2018 Skymind, Inc.
- *
- * This program and the accompanying materials are made available under the
- * terms of the Apache License, Version 2.0 which is available at
- * https://www.apache.org/licenses/LICENSE-2.0.
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
- * License for the specific language governing permissions and limitations
- * under the License.
- *
- * SPDX-License-Identifier: Apache-2.0
- ******************************************************************************/
+/*
+ *  ******************************************************************************
+ *  *
+ *  *
+ *  * This program and the accompanying materials are made available under the
+ *  * terms of the Apache License, Version 2.0 which is available at
+ *  * https://www.apache.org/licenses/LICENSE-2.0.
+ *  *
+ *  *  See the NOTICE file distributed with this work for additional
+ *  *  information regarding copyright ownership.
+ *  * Unless required by applicable law or agreed to in writing, software
+ *  * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ *  * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ *  * License for the specific language governing permissions and limitations
+ *  * under the License.
+ *  *
+ *  * SPDX-License-Identifier: Apache-2.0
+ *  *****************************************************************************
+ */
 
 package org.deeplearning4j.spark.impl.paramavg;
 
 
+import com.sun.jna.Platform;
 import org.apache.hadoop.conf.Configuration;
 import org.apache.hadoop.fs.FileSystem;
 import org.apache.hadoop.fs.LocatedFileStatus;
@@ -50,10 +55,15 @@ import org.deeplearning4j.spark.impl.graph.SparkComputationGraph;
 import org.deeplearning4j.spark.impl.multilayer.SparkDl4jMultiLayer;
 import org.deeplearning4j.spark.stats.EventStats;
 import org.deeplearning4j.spark.stats.ExampleCountEventStats;
-import org.junit.Ignore;
-import org.junit.Rule;
-import org.junit.Test;
-import org.junit.rules.TemporaryFolder;
+import org.junit.jupiter.api.Disabled;
+
+import org.junit.jupiter.api.Tag;
+import org.junit.jupiter.api.Test;
+
+import org.junit.jupiter.api.Timeout;
+import org.junit.jupiter.api.io.TempDir;
+import org.nd4j.common.tests.tags.NativeTag;
+import org.nd4j.common.tests.tags.TagNames;
 import org.nd4j.evaluation.classification.Evaluation;
 import org.nd4j.evaluation.classification.ROC;
 import org.nd4j.evaluation.classification.ROCMultiClass;
@@ -76,12 +86,12 @@ import java.io.File;
 import java.nio.file.Path;
 import java.util.*;
 
-import static org.junit.Assert.*;
+import static org.junit.jupiter.api.Assertions.*;
 
-
-/**
- * Created by agibsonccc on 1/18/15.
- */
+@Tag(TagNames.FILE_IO)
+@Tag(TagNames.SPARK)
+@Tag(TagNames.DIST_SYSTEMS)
+@NativeTag
 public class TestSparkMultiLayerParameterAveraging extends BaseSparkTest {
 
     public static class TestFn implements Function<LabeledPoint, LabeledPoint>{
@@ -91,8 +101,7 @@ public class TestSparkMultiLayerParameterAveraging extends BaseSparkTest {
         }
     }
 
-    @Rule
-    public TemporaryFolder testDir = new TemporaryFolder();
+
 
 
     @Override
@@ -112,6 +121,10 @@ public class TestSparkMultiLayerParameterAveraging extends BaseSparkTest {
 
     @Test
     public void testFromSvmLightBackprop() throws Exception {
+        if(Platform.isWindows()) {
+            //Spark tests don't run on windows
+            return;
+        }
         JavaRDD<LabeledPoint> data = MLUtils
                         .loadLibSVMFile(sc.sc(),
                                         new ClassPathResource("svmLight/iris_svmLight_0.txt").getTempFileFromArchive()
@@ -144,6 +157,10 @@ public class TestSparkMultiLayerParameterAveraging extends BaseSparkTest {
 
     @Test
     public void testFromSvmLight() throws Exception {
+        if(Platform.isWindows()) {
+            //Spark tests don't run on windows
+            return;
+        }
         JavaRDD<LabeledPoint> data = MLUtils
                         .loadLibSVMFile(sc.sc(),
                                         new ClassPathResource("svmLight/iris_svmLight_0.txt").getTempFileFromArchive()
@@ -174,7 +191,10 @@ public class TestSparkMultiLayerParameterAveraging extends BaseSparkTest {
 
     @Test
     public void testRunIteration() {
-
+        if(Platform.isWindows()) {
+            //Spark tests don't run on windows
+            return;
+        }
         DataSet dataSet = new IrisDataSetIterator(5, 5).next();
         List<DataSet> list = dataSet.asList();
         JavaRDD<DataSet> data = sc.parallelize(list);
@@ -194,6 +214,10 @@ public class TestSparkMultiLayerParameterAveraging extends BaseSparkTest {
 
     @Test
     public void testUpdaters() {
+        if(Platform.isWindows()) {
+            //Spark tests don't run on windows
+            return;
+        }
         SparkDl4jMultiLayer sparkNet = getBasicNetwork();
         MultiLayerNetwork netCopy = sparkNet.getNetwork().clone();
 
@@ -216,7 +240,10 @@ public class TestSparkMultiLayerParameterAveraging extends BaseSparkTest {
 
     @Test
     public void testEvaluation() {
-
+        if(Platform.isWindows()) {
+            //Spark tests don't run on windows
+            return;
+        }
         SparkDl4jMultiLayer sparkNet = getBasicNetwork();
         MultiLayerNetwork netCopy = sparkNet.getNetwork().clone();
 
@@ -249,7 +276,10 @@ public class TestSparkMultiLayerParameterAveraging extends BaseSparkTest {
     public void testSmallAmountOfData() {
         //Idea: Test spark training where some executors don't get any data
         //in this case: by having fewer examples (2 DataSets) than executors (local[*])
-
+        if(Platform.isWindows()) {
+            //Spark tests don't run on windows
+            return;
+        }
         MultiLayerConfiguration conf = new NeuralNetConfiguration.Builder().updater(new RmsProp())
                         .optimizationAlgo(OptimizationAlgorithm.STOCHASTIC_GRADIENT_DESCENT).list()
                         .layer(0, new org.deeplearning4j.nn.conf.layers.DenseLayer.Builder().nIn(nIn).nOut(3)
@@ -352,6 +382,10 @@ public class TestSparkMultiLayerParameterAveraging extends BaseSparkTest {
 
     @Test
     public void testParameterAveragingMultipleExamplesPerDataSet() throws Exception {
+        if(Platform.isWindows()) {
+            //Spark tests don't run on windows
+            return;
+        }
         int dataSetObjSize = 5;
         int batchSizePerExecutor = 25;
         List<DataSet> list = new ArrayList<>();
@@ -400,9 +434,13 @@ public class TestSparkMultiLayerParameterAveraging extends BaseSparkTest {
 
 
     @Test
-    public void testFitViaStringPaths() throws Exception {
-
-        Path tempDir = testDir.newFolder("DL4J-testFitViaStringPaths").toPath();
+    @Disabled("Permissions issues on CI")
+    public void testFitViaStringPaths(@TempDir Path testDir) throws Exception {
+        if(Platform.isWindows()) {
+            //Spark tests don't run on windows
+            return;
+        }
+        Path tempDir = new File(testDir.toFile(),"DL4J-testFitViaStringPaths").toPath();
         File tempDirF = tempDir.toFile();
         tempDirF.deleteOnExit();
 
@@ -464,9 +502,13 @@ public class TestSparkMultiLayerParameterAveraging extends BaseSparkTest {
     }
 
     @Test
-    public void testFitViaStringPathsSize1() throws Exception {
-
-        Path tempDir = testDir.newFolder("DL4J-testFitViaStringPathsSize1").toPath();
+    @Disabled("Permissions issues on CI")
+    public void testFitViaStringPathsSize1(@TempDir Path testDir) throws Exception {
+        if(Platform.isWindows()) {
+            //Spark tests don't run on windows
+            return;
+        }
+        Path tempDir = new File(testDir.toFile(),"DL4J-testFitViaStringPathsSize1").toPath();
         File tempDirF = tempDir.toFile();
         tempDirF.deleteOnExit();
 
@@ -545,10 +587,14 @@ public class TestSparkMultiLayerParameterAveraging extends BaseSparkTest {
 
 
     @Test
-    public void testFitViaStringPathsCompGraph() throws Exception {
-
-        Path tempDir = testDir.newFolder("DL4J-testFitViaStringPathsCG").toPath();
-        Path tempDir2 = testDir.newFolder("DL4J-testFitViaStringPathsCG-MDS").toPath();
+    @Disabled("Permissions issues on CI")
+    public void testFitViaStringPathsCompGraph(@TempDir Path testDir) throws Exception {
+        if(Platform.isWindows()) {
+            //Spark tests don't run on windows
+            return;
+        }
+        Path tempDir = new File(testDir.toFile(),"DL4J-testFitViaStringPathsCG").toPath();
+        Path tempDir2 = new File(testDir.toFile(),"DL4J-testFitViaStringPathsCG-MDS").toPath();
         File tempDirF = tempDir.toFile();
         File tempDirF2 = tempDir2.toFile();
         tempDirF.deleteOnExit();
@@ -640,9 +686,12 @@ public class TestSparkMultiLayerParameterAveraging extends BaseSparkTest {
 
 
     @Test
-    @Ignore("AB 2019/05/23 - Failing on CI only - passing locally. Possible precision or threading issue")
+    @Disabled("AB 2019/05/23 - Failing on CI only - passing locally. Possible precision or threading issue")
     public void testSeedRepeatability() throws Exception {
-
+        if(Platform.isWindows()) {
+            //Spark tests don't run on windows
+            return;
+        }
         MultiLayerConfiguration conf = new NeuralNetConfiguration.Builder().seed(12345).updater(new RmsProp())
                         .optimizationAlgo(OptimizationAlgorithm.STOCHASTIC_GRADIENT_DESCENT)
                         .weightInit(WeightInit.XAVIER).list()
@@ -707,13 +756,17 @@ public class TestSparkMultiLayerParameterAveraging extends BaseSparkTest {
 
         boolean eq1 = p1.equalsWithEps(p2, 0.01);
         boolean eq2 = p1.equalsWithEps(p3, 0.01);
-        assertTrue("Model 1 and 2 params should be equal", eq1);
-        assertFalse("Model 1 and 3 params shoud be different", eq2);
+        assertTrue(eq1, "Model 1 and 2 params should be equal");
+        assertFalse(eq2, "Model 1 and 3 params shoud be different");
     }
 
 
     @Test
     public void testIterationCounts() throws Exception {
+        if(Platform.isWindows()) {
+            //Spark tests don't run on windows
+            return;
+        }
         int dataSetObjSize = 5;
         int batchSizePerExecutor = 25;
         List<DataSet> list = new ArrayList<>();
@@ -760,6 +813,10 @@ public class TestSparkMultiLayerParameterAveraging extends BaseSparkTest {
 
     @Test
     public void testIterationCountsGraph() throws Exception {
+        if(Platform.isWindows()) {
+            //Spark tests don't run on windows
+            return;
+        }
         int dataSetObjSize = 5;
         int batchSizePerExecutor = 25;
         List<DataSet> list = new ArrayList<>();
@@ -805,7 +862,7 @@ public class TestSparkMultiLayerParameterAveraging extends BaseSparkTest {
 
 
     @Test
-    @Ignore   //Ignored 2019/04/09 - low priority: https://github.com/deeplearning4j/deeplearning4j/issues/6656
+    @Disabled   //Ignored 2019/04/09 - low priority: https://github.com/eclipse/deeplearning4j/issues/6656
     public void testVaePretrainSimple() {
         //Simple sanity check on pretraining
         int nIn = 8;
@@ -841,7 +898,7 @@ public class TestSparkMultiLayerParameterAveraging extends BaseSparkTest {
     }
 
     @Test
-    @Ignore    //Ignored 2019/04/09 - low priority: https://github.com/deeplearning4j/deeplearning4j/issues/6656
+    @Disabled    //Ignored 2019/04/09 - low priority: https://github.com/eclipse/deeplearning4j/issues/6656
     public void testVaePretrainSimpleCG() {
         //Simple sanity check on pretraining
         int nIn = 8;
@@ -989,9 +1046,13 @@ public class TestSparkMultiLayerParameterAveraging extends BaseSparkTest {
     }
 
 
-    @Test(timeout = 120000L)
+    @Test()
+    @Timeout(120000)
     public void testEpochCounter() throws Exception {
-
+        if(Platform.isWindows()) {
+            //Spark tests don't run on windows
+            return;
+        }
         MultiLayerConfiguration conf = new NeuralNetConfiguration.Builder()
                 .list()
                 .layer(new OutputLayer.Builder().nIn(4).nOut(3).build())
